requirements.txt
numpy
pandas
wget
scipy
networkx
torch>=1.12
matplotlib
transformers==4.38.2
seaborn
tqdm
scikit-learn
pytorch-lightning==1.9.0
torchmetrics
biopython
pandarallel